fix: statistic,review,order & dashboard

+++ src/main/java/com/gelitix/backend/dashboard/dto/EventStatisticsDTO.java
@@ -1,6 +1,5 @@
 package com.gelitix.backend.dashboard.dto;
 
-import com.gelitix.backend.order.dao.EventAttendeeCountDao;
 import com.gelitix.backend.order.dao.PeriodicalRevenueDao;
 import lombok.Data;
 
@@ -10,8 +9,8 @@ import java.util.List;
 @Data
 public class EventStatisticsDTO {
     private BigDecimal totalRevenue;
-    private Double totalOrders;
-    private Double totalCustomers;
+    private Long totalOrders;
+    private Long totalCustomers;
     private List<PeriodicalRevenueDao> yearlyRevenue;
     private List<PeriodicalRevenueDao> monthlyRevenue;
     private List<PeriodicalRevenueDao> dailyRevenue;
